pyproject.toml
[project]
name = "venv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]
